Changed the health_, attackPower_, defense_ and its setter/getter functions to type double. I made this change because I am multiplying by decimals and require more accurate representation for these data members.

+++ RPG/Character.cpp
@@ -1,20 +1,20 @@
 #include "Character.h"
 
 // Constructor
-Character::Character(int attackPower, int defense, const std::string& name) : attackPower_(attackPower), defense_(defense), name_(name) {}
+Character::Character(double attackPower, double defense, const std::string& name) : attackPower_(attackPower), defense_(defense), name_(name) {}
 
 // Setter Methods
-void Character::setAttackPower(int attackPower) { attackPower_ = attackPower; }
-void Character::setDefenseLevel(int defense) { defense_ = defense; }
-void Character::setHealth(int health) { health_ = health; }
+void Character::setAttackPower(double attackPower) { attackPower_ = attackPower; }
+void Character::setDefenseLevel(double defense) { defense_ = defense; }
+void Character::setHealth(double health) { health_ = health; }
 void Character::setCombatLevel(int combatLevel) { combatLevel_ = combatLevel; }
 void Character::setExperience(int experience) { totalExperience_ = experience; }
 void Character::setName(const std::string& name) { name_ = name; }
 
 // Getter Methods
-int Character::getAttackPower() const { return attackPower_; }
-int Character::getDefenseLevel() const { return defense_; }
-int Character::getHealth() const { return health_; }
+double Character::getAttackPower() const { return attackPower_; }
+double Character::getDefenseLevel() const { return defense_; }
+double Character::getHealth() const { return health_; }
 int Character::getCombatLevel() const { return combatLevel_; }
 int Character::getExperience() const { return totalExperience_; }
 std::string Character::getName() const { return name_; }
